+++ docs/releases.md
@@ -3,6 +3,12 @@
 This page gathered the changes made between version of the launchpy module.\
 This has been started after the 0.3.0 release.\
 
+## 0.4.4
+* adding a `getBuilds` method on the `Library` class instance
+* extend support for pagination on `getEnvironments`
+* adding `getProductionEndpoint` and `getStagingEndpoint` to retrieve the URL to use on websites
+* support the instantiation of the `Library` class via a single libraryId string. 
+
 ## 0.4.3
 * Refactor `Translator` class for removing `pandas` dependency\
 Patch: